feat(s2n-quic-core): add state contstruction
